add more safe_enumerate
